Close database connection

This method is an adaptation to the INBO databases from the eponymous function in the odbc package and is an implementation of the method DBI::dbDisconnect() defined in the DBI package.

Usage

# S4 method for class 'OdbcConnection'
dbDisconnect(conn, ...)

Arguments

conn

A DBI::DBIConnection object, as returned by connect_inbo_dbase().

...

Other parameters passed on to methods.

See also