Get unique combinations of taxon, IFBL-square and year.

This functions queries all validated observations of the florabank database and returns unique combinations of taxon, IFBL-square and year. Either a 1 km by 1 km or a 4 km x 4 km resolution can be chosen and a begin year can be set. Observations of taxa at genus level or higher are excluded. The taxonomic group can be chosen.

Usage

get_florabank_taxon_ifbl_year(
  connection,
  starting_year = 2010,
  ifbl_resolution = c("1km-by-1km", "4km-by-4km"),
  taxongroup = c("Vaatplanten", "Mossen", "Lichenen (korstmossen)", "Kranswieren"),
  collect = FALSE
)

Arguments

connection

A connection to the florabank database. See the example section for how to connect and disconnect to the database.

starting_year

Filter for observations that start from this year onwards. Default is 2010.

ifbl_resolution

The requested spatial resolution can be either 1km-by-1km IFBL squares or 4km-by-4km. Default is 1km-by-1km.

taxongroup

Choose for which taxonomic group you want the unique combinations. One of "Vaatplanten" (the default), "Mossen", "Lichenen (korstmossen)" or "Kranswieren".

collect

If FALSE (the default), a remote tbl object is returned. This is like a reference to the result of the query but the full result of the query is not brought into memory. If TRUE the full result of the query is collected (fetched) from the database and brought into memory of the working environment.

Value

A dataframe with one line for each combination of taxon, IFBL-square (either at 1 km x 1 km or 4 km x 4 km resolution) and year. In case the resolution is 1 km x 1 km, a variable ifbl_4by4 gives the corresponding 4 km x 4 km square within which the 1 km x 1 km square is located. In case the resolution is 4 km x 4 km the variable ifbl_number_squares gives the number of unique nested squares where the taxon was observed for that year and 4 x 4 square combination.

See also

Examples

if (FALSE) { # \dontrun{
library(inbodb)
# connect to florabank
db_connectie <- connect_inbo_dbase("D0152_00_Flora")

# get records at 1 km x 1 km resolution for vascular plants from 2010
# (default) without collecting all data into memory (default).
fb_kwartier <- get_florabank_taxon_ifbl_year(db_connectie)
# to collect the data in memory set collect to TRUE or do
fb_kwartier <- dplyr::collect(fb_kwartier)

# get records at 4 km x 4 km resolution starting from 2000
fb_uur <- get_florabank_taxon_ifbl_year(db_connectie, starting_year = 2000,
 ifbl_resolution = "4km-by-4km", taxongroup = "Mossen")

# disconnect from florabank
dbDisconnect(db_connectie)
} # }